update tailwindcss and write fontend and update api post and question

+++ server/src/controllers/post.js
@@ -1,11 +1,24 @@
 const Post = require("../models/post");
+const User = require("../models/user");
 const asyncHandler = require("express-async-handler");
 
 const createPost = asyncHandler(async (req, res) => {
-  const { title, content, state } = req.body;
+  const { title, content, state, tags } = req.body;
+  // const { _id } = req.user;
   console.log(title, content);
-  if (!title || !content) throw new Error("missing input");
-  const response = await Post.create({ title, content });
+  // if (!_id || !title || !content) throw new Error("missing input");
+  if (!title || !content || !tags) throw new Error("missing input");
+  const response = await Post.create({ title, content, state, tags });
+  // const response = await Post.create({ title, content, state, idAuthor: _id });
+
+  // if (response) {
+  //   // Update the user to add the post ID to the posts array
+  //   await User.findByIdAndUpdate(
+  //     _id,
+  //     { $push: { posts: response._id } },
+  //     { new: true }
+  //   );
+  // }
   return res.json({
     success: response ? true : false,
     createPost: response ? response : "cannot create new post",
@@ -137,10 +150,18 @@ const getPost = asyncHandler(async (req, res) => {
 
 const deletePost = asyncHandler(async (req, res) => {
   const { pid } = req.params;
-  const post = await Post.findByIdAndDelete(pid);
+  const { _id } = req.user;
+  const post = await Post.findById(pid);
+  if (!post) throw new Error("Post not found");
+  const response = await Post.findByIdAndDelete(pid);
+
+  if (response) {
+    // Update the user to remove the post ID from the posts array
+    await User.findByIdAndUpdate(_id, { $pull: { posts: pid } }, { new: true });
+  }
   return res.json({
-    success: post ? true : false,
-    deletedPost: post || "something went wrong",
+    success: response ? true : false,
+    deletedPost: response ? response : "cannot delete post",
   });
 });
 
@@ -153,7 +174,264 @@ const deletePost = asyncHandler(async (req, res) => {
 //       updatedBlog: response ? response : 'cannot upload image for blog'
 //   })
 // })
+const commentPost = asyncHandler(async (req, res) => {
+  const { _id } = req.user;
+  const { pid } = req.params;
+  const { comment } = req.body;
+  if (!pid || !comment) throw new Error("Missing inputs");
+  const post = await Post.findById(pid);
+  if (!post) throw new Error("Post not found!");
 
+  if (comment.trim().length === 0 || comment.length > 500) {
+    return res
+      .status(400)
+      .json({ success: false, message: "Comment cannot be empty or too long" });
+  }
+  const response = await Post.findByIdAndUpdate(
+    pid,
+    { $push: { comments: { postedBy: _id, comment: comment } } },
+    { new: true }
+  );
+
+  return res.json({
+    success: response ? true : false,
+    rs: response,
+  });
+});
+
+const repliesPost = asyncHandler(async (req, res) => {
+  const { _id } = req.user;
+  const { pid, cid } = req.params;
+  const { comment } = req.body;
+
+  if (!pid || !cid || !_id || !comment) throw new Error("Missing in put");
+
+  const post = await Post.findById(pid);
+  if (!post) throw new Error("Post not found");
+
+  if (comment.trim().length === 0 || comment.length > 500) {
+    return res
+      .status(400)
+      .json({ success: false, message: "Comment cannot be empty or too long" });
+  }
+  // Find the comment
+  const commentIndex = post.comments.findIndex((c) => c._id.toString() === cid);
+  if (commentIndex === -1) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Comment not found" });
+  }
+
+  // Add the reply to the comment
+  post.comments[commentIndex].replies.push({ postedBy: _id, comment: comment });
+
+  // Save the post
+  const updatedPost = await post.save();
+
+  return res.status(200).json({
+    success: true,
+    message: "Reply added successfully",
+    post: updatedPost,
+  });
+});
+
+const deleteComment = asyncHandler(async (req, res) => {
+  const { _id } = req.user;
+  const { cid, pid } = req.params;
+
+  if (!pid || !cid || !_id) throw new Error("Missing in put");
+
+  const post = await Post.findById(pid);
+  // Find the comment
+  const commentIndex = post.comments.findIndex((c) => c._id.toString() === cid);
+  if (commentIndex === -1) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Comment not found" });
+  }
+
+  // Add the reply to the comment
+  if (_id.toString() !== post.comments[commentIndex].postedBy.toString())
+    throw new Error("You can not delete comment of another user");
+
+  // Remove the comment
+  post.comments.splice(commentIndex, 1);
+
+  // Save the post
+  const deletedPost = await post.save();
+  return res.status(200).json({
+    success: true,
+    message: "Comment deleted successfully",
+    post: deletedPost,
+  });
+});
+
+const deleteReplies = asyncHandler(async (req, res) => {
+  const { _id } = req.user;
+  const { pid, cid, rid } = req.params;
+  if (!pid || !cid || !_id || !rid) throw new Error("Missing in put");
+
+  const post = await Post.findById(pid);
+  if (!post) throw new Error("Post not found");
+  // Find the comment
+  const commentIndex = post.comments.findIndex((c) => c._id.toString() === cid);
+  if (commentIndex === -1) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Comment not found" });
+  }
+  const replyIndex = post.comments[commentIndex].replies.findIndex(
+    (c) => c._id.toString() === rid
+  );
+  if (replyIndex === -1) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Replies not found" });
+  }
+
+  // Add the reply to the comment
+  if (
+    _id.toString() !==
+    post.comments[commentIndex].replies[replyIndex].postedBy.toString()
+  )
+    throw new Error("You can not delete comment of another user");
+
+  // Remove the comment
+  post.comments[commentIndex].replies.splice(replyIndex, 1);
+
+  // Save the post
+  const deletedPost = await post.save();
+  return res.status(200).json({
+    success: true,
+    message: "Comment deleted successfully",
+    post: deletedPost,
+  });
+});
+
+const updateComment = asyncHandler(async (req, res) => {
+  const { _id } = req.user;
+  const { pid, cid } = req.params;
+  const { comment } = req.body;
+  if (!pid || !cid || !comment || !_id) throw new Error("Missing input");
+  if (comment.trim().length === 0 || comment.length > 500) {
+    return res
+      .status(400)
+      .json({ success: false, message: "Comment cannot be empty or too long" });
+  }
+  const post = await Post.findById(pid);
+  if (!post) throw new Error("Post not found");
+  // Find the comment
+  const commentIndex = post.comments.findIndex((c) => c._id.toString() === cid);
+  if (commentIndex === -1) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Comment not found" });
+  }
+
+  // Check if the user is the owner of the comment
+  if (_id.toString() !== post.comments[commentIndex].postedBy.toString()) {
+    return res.status(403).json({
+      success: false,
+      message: "You cannot update another user's comment",
+    });
+  }
+
+  // Update the comment
+  post.comments[commentIndex].comment = comment;
+
+  // Save the post
+  const updatedPost = await post.save();
+
+  return res.status(200).json({
+    success: true,
+    message: "Comment updated successfully",
+    post: updatedPost,
+  });
+});
+
+const updateReply = asyncHandler(async (req, res) => {
+  const { _id } = req.user;
+  const { pid, cid, rid } = req.params;
+  const { comment } = req.body;
+
+  // Validate inputs
+  if (!pid || !cid || !rid || !_id || !comment)
+    return res.status(400).json({ success: false, message: "Missing input" });
+
+  if (comment.trim().length === 0 || comment.length > 500)
+    throw new Error("Comment cannot be empty or too long");
+
+  // Find the post by id
+  const post = await Post.findById(pid);
+  if (!post) throw new Error("Post not found");
+
+  // Find the comment within the post
+  const commentIndex = post.comments.findIndex((c) => c._id.toString() === cid);
+  if (commentIndex === -1) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Comment not found" });
+  }
+
+  const commentObj = post.comments[commentIndex];
+
+  // Find the reply within the comment
+  const replyIndex = commentObj.replies.findIndex(
+    (r) => r._id.toString() === rid
+  );
+  if (replyIndex === -1) {
+    return res.status(404).json({ success: false, message: "Reply not found" });
+  }
+
+  const replyObj = commentObj.replies[replyIndex];
+
+  // Check if the user is the owner of the reply
+  if (_id.toString() !== replyObj.postedBy.toString()) {
+    return res.status(403).json({
+      success: false,
+      message: "You cannot update another user's reply",
+    });
+  }
+
+  // Update the reply
+  replyObj.comment = comment;
+
+  // Save the updated post
+  const updatedPost = await post.save();
+
+  return res.status(200).json({
+    success: true,
+    message: "Reply updated successfully",
+    post: updatedPost,
+  });
+});
+
+const addPostInFavorites = asyncHandler(async (req, res) => {
+  const { pid } = req.params;
+  const { _id } = req.user;
+  if (!_id || !pid) throw new Error("missing input");
+  // Tạo đối tượng mới để thêm vào listfavorite
+  const newFavorite = {
+    item: pid, // item là pid (ID của bài viết)
+    itemType: "Post", // Xác định loại là 'Post'
+  };
+
+  // Tìm user theo _id và đẩy mục yêu thích mới vào mảng listfavorite
+  const response = await User.findByIdAndUpdate(
+    _id,
+    { $push: { listfavorite: newFavorite } }, // Sử dụng $push để thêm mục yêu thích
+    { new: true } // Trả về document mới sau khi cập nhật
+  );
+
+  return res.json({
+    success: response ? true : false,
+    createPost: response ? response : "cannot add to favorites",
+  });
+});
+
+//check create post
+//check delete post
+//check addpostInFavorite
 module.exports = {
   createPost,
   updatePost,
@@ -161,4 +439,12 @@ module.exports = {
   likePost,
   dislikePost,
   getPosts,
+  deletePost,
+  commentPost,
+  repliesPost,
+  deleteComment,
+  deleteReplies,
+  updateComment,
+  updateReply,
+  addPostInFavorites,
 };
